+++ src/manifest.ts
@@ -206,6 +206,7 @@ export interface SentenceCasePluginConfig extends ConfigurablePluginType {
 }
 export interface WorkspacePluginConfig extends ConfigurablePluginType {
   merge?: boolean;
+  considerAllArtifacts?: boolean;
 }
 export interface GroupPriorityPluginConfig extends ConfigurablePluginType {
   groups: string[];